+++ CHANGELOG.md
@@ -1,4 +1,16 @@
-5.6
-----
+5.7 (2014/10/10)
+----------------
+
+Add WebimUser, WebimVisitor, WebimRobot
+
+Add WebimCluster
+
+Modify WebimClient to support cluster
+
+Fix author comments
+
+
+5.6 (2014/09/10)
+--------------------
 
 WebimClient.java: List<String> -> Set<String> to avoid duplicate keys